requirements.txt
tweepy
pandas
json
apache-airflow
cassandra-sigv4
cassandra-driver
apache-airflow-providers-apache-cassandra